Fix #5: compatibility with merlin
Reported by Kiran Gopinathan,
https://github.com/thierry-martinez/metapp/issues/5

Merlin runs preprocessor in a reduced environment where `PATH` only
contains the directory where ocaml is (typically
`~/.opam/<switch>/bin`). This prevents ocaml from finding the
assembler (`as`).

This commit makes metapp check whether `as` is in the PATH before
calling ocaml. If it is not, it checks whether `/usr/bin/as` exists.
It it exists, `/usr/bin` is appended to PATH.

This is still quite fragile and will not work in environments where
`as` is not in `/usr/bin/`, but I do not know how to do better.